:sparkles: build spring cli tools
